requirements.txt
python-dotenv
langchain_openai
langchain_chroma
langchain_community
langgraph
langgraph.prebuilt
langchain-docling